+++ GT-Buses/GBImage.h
@@ -13,8 +13,8 @@ extern float const kRouteImageViewSize;
 @class GBStopGroup, GBRoute;
 @interface UIImage (GBImage)
 
-+ (UIImage *)arrowImageWithColor:(UIColor *)color size:(CGSize)size;
-+ (UIImage *)circleStopImageWithColor:(UIColor *)color size:(float)size;
++ (UIImage *)arrowImageWithColor:(UIColor *)color;
++ (UIImage *)circleStopImageWithColor:(UIColor *)color;
 + (UIImage *)circlesWithStopGroup:(GBStopGroup *)stopGroup;
 + (UIImage *)circleRouteImageWithRoute:(GBRoute *)route;